1. Contains output partial vectors in contiguous memory locations unlike previous commit which had it in pthread malloc space separately. 2. Added a parallel reduce method for the output elements for the manycore version of algo. Note that earlier we did serial output but each element parallely.
Result: Parallel output reduce seems to work 5x faster and takes less than 10%
of algo time. Should be similar for simd version since we do reduce in manycore
version. Only difference will be that we figure out data to be taken from
active vector cores only (not scalar cores, not the ones which were not assigned
in template).