Use naming convention, not func name CLI arg
We now transform any function called `tril_*`, rather than requiring a
command-line argument to identify the kernel function. This will help
with multi-kernel support because we absolutely do not want to be
specifying a million different kernel names on the CLI.

This also adapts the `simple` test, but there are more to be fixed.