ATAX: use common Makefile
Convert the ATAX benchmark to use the common Makefile, just like vvadd.

To make this work, this also adds a couple of extra Make variables to the
common Makefile: `GEM5_ARGS` and `HB_ARGS`. Because they are assigned with
`?=`, they can be overridden by the "invoking" Makefiles by just assigning to
them.

Change-Id: Ie6f5f5235288ba92adadcaf926d8a0530cd0e03e