Refactor: centralize kernel state initialization
This undoes the ugly copy-paste junk I introduced in #36 to
re-initialize the parser state when a kernel finished *and* at the top
of the file. Now we only do it once, for the most part: when *starting*
a kernel function.

(This will make it a little easier to change the "auxiliary block" label
in only one place.)