Add `end at_jump` delimiter
This is an alternative to using `until_next`. You use `begin` and `end`,
but you qualify the ending delimiter like `end at_jump`. This *almost*
ends the block, but not quite: it keeps gobbling up a few more
instructions *until it hits a jump*. This copes with cases where a few
instructions are moved *out* of a begin/end-delimited block at the *end*
of a loop, and you want to recover those instructions. But you *don't*
want `until_next`, which will keep going and reach "outside" of the
loop.

With this, I think the recommendation should now be that you should
*always* use `if_begin`; I don't yet know of a situation where that's
unsafe.