Support "return" blocks in any order
This actually does something a little more subtle: it moves the
"function wrap-up" stuff in the vector-code extractor. It used to happen
right when we parsed the return block. Now it happens when the function
actually finishes (as indicated by `.size`). So in general, now blocks
can happen in any order.