Move scalar cleanup code to just before devec
The problem, as reported by @neiladit, was that GCC would move around
the basic block for the `devec` instruction. Then the cleanup code being
pasted just *above* that block's label wouldn't get run when jumping to
that label. Now we put the code after that label.

As a consequence, we have to generate a new label for the `devec`
instruction, which requires its own label. We do this by creating a new
label name based on the existing label name used in the `devec`
instruction. Because we invented the label, it's impossible that anyone
else is jumping to this location.